worked on confusion based on cirikov map from paper. focused on 768 by 768 singular frame. was able to produce a correct frame of the confused image (afterConfusion.png) after confrontation of serialized python version in my own pc. also built inverseconfusionkernel to verify for sure that the confusedImage could map back to original given frame (afterInvConfusion.png)

+++ include/prbg_main_plcm.hpp
@@ -4,6 +4,6 @@
 #include <vector>
 #include <iostream>
 
-std::vector<double> generatePRBGMainKeys(double x0, double p, int numParameters4subsequentPRBGas, double *sc);
+std::vector<double> generatePRBGMainKeys(double x0, double p, int numParameters4subsequentPRBGas, int *sc);
 
 #endif
